Supports mapping between MOSART and RDycore
MOSART/Rtm domain can now be of a different size as compared to RDycore.
In such a configuration, a nearest neighbor map is used to map the data
from MOSART/Rtm domain onto RDycore domain. The map is a PETSc Vec in
binary format that should be named `map_MOSART_to_RDycore.bin` and
present in the run directory.